rewrote pss-system
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@243 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/global_full/global_control.c
@@ -56,9 +56,9 @@ extern struct _FILES  allfiles;
  *----------------------------------------------------------------------*/
 void ntam(int argc, char *argv[])
 {
-int ierr;
-int handle;
-double t0,ti,tc;
+int      ierr;
+long int handle;
+double   t0,ti,tc;
 /*---------------------------------------- init devices, tracing, etc...*/
 ntaini(argc,argv);
 /*--------------------------------input phase, input of all information */
